Доработана логика всех 3х сенсоров, Chi заполняется полностью.

+++ api/src/main/java/entities/devices/cpus/CpuCore.java
@@ -8,5 +8,5 @@ public class CpuCore {
     @JsonProperty("coreName")
     private String coreName;
     @JsonProperty("temperature")
-    private long temperature;
+    private double temperature;
 }
